refactor: 🔥 rewrite to support non-creature objects
No non-creature objects were addedd, but this split apart the serialization steps to be traits which lets more object parsing be added in the future with out a ton of trouble. We now check the object type of the raw file first, then go and parse it. It adds an extra file read step (unfortunately). It may be possible to streamline further by opening file etc before checking and then parsring.

Some library method signatures were modified